refactor(criteria): split the reviewed workflow grammars into their own module
The file-size gate went red on main when #225 merged: the criterion
machine was 899 lines at base, one under the 900-line fail threshold, and
the destination-tail fix pushed it to 921. Split, never allowlist: the
four SELF-CONTAINED reviewed workflow/copy grammars (strategy-board
allocation, screen-gate workflow, server-rendered offer-audience
description/summary, named-lender rate query) and their two public
helpers move to marketing_selection_reviewed_workflows; the
consent-reroute grammar stays because it embeds the coreference
population fragment. Criteria is 853 lines; the new module is 91.

Pure move, proven: the #225 battery rerun post-split is verdict-identical
(1,482 probes x four surfaces, 0 diffs, distinct file hashes on each
side), the exact CI size-gate invocation exits 0, and ruff + mypy + the
full suite are green with the moved helpers' round23 pins re-pointed.
